More IWYU and header dependency cycle breaking
The primary intent of this commit was to remove options.h from the
dependencies of beam.h, since it was only ever used to set a single
struct field's default value. That can be done perfectly well with a
constructor in a cc file.

Of course, removing options.h from beam.h broke all kinds of other
things that had come to depend on options.h being included from their
dependencies, or, to be precise, on *dependencies* of options.h (mostly
mpr.h in practice). What a mess.

This change also adds five more headers to the set of headers that are
buildable: i.e. they (transitively) include all their own dependencies.